[ADD] #5: Moved journey modes from MainActivity.kt to JourneyModesFragment.kt
